+++ src/test/java/calculator/CalculatorTest.java
@@ -1,6 +1,5 @@
-package CalculatorTest;
+package calculator;
 
-import Calculator.Calculator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
